refactor: Paging 훅 유틸함수로 변경
