+++ src/components/postDetail/ReviewList/index.tsx
@@ -3,14 +3,13 @@ import { useState } from 'react';
 import classNames from 'classnames/bind';
 
 import { REVIEW_SORT_OPTIONS } from '@/constants';
-import { getSortedDataList, getReviewPageSize } from '@/utils';
+import { getSortedDataList, getReviewPageSize, getPaginatedDataList } from '@/utils';
 
 import Dropdown from '@/components/commons/Dropdown';
 import Pagination from '@/components/commons/Pagination';
 import ReviewCard from '@/components/commons/ReviewCard';
 import ReviewSummary from '@/components/postDetail/ReviewSummary';
 import { useDeviceType } from '@/hooks/useDeviceType';
-import usePaginatedDataList from '@/hooks/usePaginatedDataList';
 
 import { Order, Review, ReviewResponse, SortOption } from '@/types';
 
@@ -45,10 +44,9 @@ const ReviewList = ({ list, profileImageUrl, nickname, email }: ReviewListProps)
     sortOption,
   });
 
-  const pagedDataList = usePaginatedDataList({
+  const pagedDataList = getPaginatedDataList({
     initialDataList: sortedDataList,
     page,
-    setPage,
     postsPerPage: pageSize,
   });
