+++ src/utils/index.ts
@@ -9,6 +9,7 @@ export * from './postFormUtils';
 export * from './isValidGameName';
 export * from './checkAuth';
 export * from './cookieUtils';
+export * from './getPageSize';
 export * from './getFilteredDataList';
 export * from './getSortedDataList';
-export * from './getPageSize';
+export * from './getPaginatedDataList';
